# swagger.json
{
    "swagger" : "2.0",
    "info" : {
      "description" : "API per l'accesso diretto ad alcune funzioni di Agrismart",
      "version" : "1.0.0",
      "title" : "Agrismart API",
      "contact" : {
        "email" : "agrismart@agrismart.com"
      },
      "license" : {
        "name" : "Apache 2.0",
        "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
      }
    },
    "host" : "localhost:8888",
    "basePath" : "/",
    "schemes" : [ "http" ],
    "paths" : {
      "/api/users/{userId}/campi" : {
        "get" : {
          "summary" : "Ritorna la lista dei campi dato un utente",
          "parameters" : [ {
            "name" : "userId",
            "in" : "path",
            "description" : "id dell'utente",
            "required" : true,
            "type" : "string"
          },
          {
            "name" : "token",
            "in": "query",
            "description" : "token dell'utente",
            "required" : true,
            "type" : "string"
          } ],
          "responses" : {
            "200" : {
              "description" : "OK"
            },
            "401" : {
              "description" : "UNAUTHORIZED"
            }
          }
        }
      },
      "/api/users/{userId}/campi/{campoId}/sensors" : {
        "get" : {
          "summary" : "Ritorna la lista dei sensori dato un campo",
          "parameters" : [ {
            "name" : "userId",
            "in" : "path",
            "description" : "id dell'utente",
            "required" : true,
            "type" : "string"
          }, {
            "name" : "campoId",
            "in" : "path",
            "description" : "è l'id del campo",
            "required" : true,
            "type" : "string"
          },
          {
            "name" : "token",
            "in": "query",
            "description" : "token dell'utente",
            "required" : true,
            "type" : "string"
          } ],
          "responses" : {
            "200" : {
              "description" : "OK"
            },
            "401" : {
              "description" : "UNAUTHORIZED"
            }
          }
        }
      },
      "/api/users/{userId}/campi/{campoId}/sensors/{sensorId}/rilevazioni" : {
        "get" : {
          "summary" : "Ritorna la lista le rilevazioni dato un sensore",
          "parameters" : [ {
            "name" : "userId",
            "in" : "path",
            "description" : "id dell'utente",
            "required" : true,
            "type" : "string"
          }, {
            "name" : "campoId",
            "in" : "path",
            "description" : "è l'id del campo",
            "required" : true,
            "type" : "string"
          }, {
            "name" : "sensorId",
            "in" : "path",
            "description" : "id del sensore",
            "required" : true,
            "type" : "string"
          },
          {
            "name" : "token",
            "in": "query",
            "description" : "token dell'utente",
            "required" : true,
            "type" : "string"
          } ],
          "responses" : {
            "200" : {
              "description" : "OK"
            },
            "401" : {
              "description" : "UNAUTHORIZED"
            }
          }
        }
      },
      "/api/meteo/{città}" : {
        "get" : {
          "summary" : "Ritorna il meteo data una città",
          "parameters" : [ {
            "name" : "città",
            "in" : "path",
            "description" : "nome della città",
            "required" : true,
            "type" : "string"
          },
          {
            "name" : "token",
            "in": "query",
            "description" : "token dell'utente",
            "required" : true,
            "type" : "string"
          } ],
          "responses" : {
            "200" : {
              "description" : "OK"
            },
            "401" : {
              "description" : "UNAUTHORIZED"
            }
          }
        }
      },
      "/api/users/{userId}/campi/{campoId}/sensors/{sensorId}/stato" : {
        "get" : {
          "summary" : "Ritorna lo stato di un sensore dato un utente, un campo ed un sensore",
          "parameters" : [ {
            "name" : "userId",
            "in" : "path",
            "description" : "id dell'utente",
            "required" : true,
            "type" : "string"
          }, {
            "name" : "sensorId",
            "in" : "path",
            "description" : "id del sensore",
            "required" : true,
            "type" : "string"
          },
        {
            "name" : "token",
            "in": "query",
            "description" : "token dell'utente",
            "required" : true,
            "type" : "string"
          } ],
          "responses" : {
            "200" : {
              "description" : "OK"
            },
            "401" : {
              "description" : "UNAUTHORIZED"
            }
          }
        }
      }
    }
  }
  